formalize the persistables container
