+++ src/Factory.php
@@ -511,23 +511,7 @@ class Factory
         }
         $collection = null;
         if (count($descendants)) {
-            $collection = new class ([]) extends Collection {
-                public function getEntityClass(): string
-                {
-                    return Collection::class;
-                }
-
-                protected function validate(mixed $value): void
-                {
-                    if (!($value instanceof Collection || $value instanceof Persistable)) {
-                        throw new InvalidArgumentException(sprintf(
-                            'Value must be an instance of %s or %s',
-                            Collection::class,
-                            Persistable::class
-                        ));
-                    }
-                }
-            };
+            $collection = new Container();
             foreach ($descendants as $property => $descendant) {
                 $collection->set($property, $descendant);
             }
